# Bitwise speed up for execution in a non vectorized environments