--- a/src/Nethermind/Nethermind.Benchmark/Evm/BitwiseAndBenchmark.cs
+++ b/src/Nethermind/Nethermind.Benchmark/Evm/BitwiseAndBenchmark.cs
@@ -18,6 +18,8 @@
 
 using System;
 using System.Numerics;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 using BenchmarkDotNet.Attributes;
 
 namespace Nethermind.Benchmarks.Evm
@@ -40,10 +42,14 @@ namespace Nethermind.Benchmarks.Evm
         [Benchmark(Baseline = true)]
         public void Current()
         {
-            for (int i = 0; i < 32; i++)
-            {
-                c[i] = (byte)(a[i] & b[i]);
-            }
+            ref var refA = ref MemoryMarshal.AsRef<ulong>(a);
+            ref var refB = ref MemoryMarshal.AsRef<ulong>(b);
+            ref var refBuffer = ref MemoryMarshal.AsRef<ulong>(c);
+
+            refBuffer = refA & refB;
+            Unsafe.Add(ref refBuffer, 1) = Unsafe.Add(ref refA, 1) & Unsafe.Add(ref refB, 1);
+            Unsafe.Add(ref refBuffer, 2) = Unsafe.Add(ref refA, 2) & Unsafe.Add(ref refB, 2);
+            Unsafe.Add(ref refBuffer, 3) = Unsafe.Add(ref refA, 3) & Unsafe.Add(ref refB, 3);
         }
         
         [Benchmark]