This PR provides an improvement over bitwise operations for non SIMD-enabled executions. It shows a tremendous gain for non-vectorized (up to 10-15 times faster).

### Details

It does it by treating `Span<byte>` as `ref ulong` and then using ref arithmetic and the fact that the size of the span is known up front, to batch computation and skip boundary checks.

### BitwiseAndBenchmark

Before

|   Method |       Mean |     Error |    StdDev | Ratio | Allocated |
|--------- |-----------:|----------:|----------:|------:|----------:|
|  Current | 38.3080 ns | 0.8547 ns | 2.1285 ns |  1.00 |         - |
| Improved |  0.7642 ns | 0.0574 ns | 0.1237 ns |  0.02 |         - |

After

|   Method |      Mean |     Error |    StdDev | Ratio | RatioSD | Allocated |
|--------- |----------:|----------:|----------:|------:|--------:|----------:|
|  Current | 2.7912 ns | 0.0882 ns | 0.1498 ns |  1.00 |    0.00 |         - |
| Improved | 0.8369 ns | 0.0587 ns | 0.1103 ns |  0.30 |    0.04 |         - |

### BitwiseNotBenchmark

Before

|   Method |       Mean |     Error |    StdDev | Ratio |  Allocated |
|--------- |-----------:|----------:|----------:|------:|------:|
|  Current | 34.7140 ns | 0.7003 ns | 0.8857 ns |  1.00 |          - |
| Improved |  0.7860 ns | 0.0562 ns | 0.0859 ns |  0.02 |         - |

After

|   Method |     Mean |     Error |    StdDev | Ratio | RatioSD | Allocated |
|--------- |---------:|----------:|----------:|------:|--------:|----------:|
|  Current | 1.578 ns | 0.0707 ns | 0.0757 ns |  1.00 |    0.00 |         - |
| Improved | 1.030 ns | 0.0611 ns | 0.1086 ns |  0.66 |    0.07 |         - |

### BitwiseOrBenchmark

Before

|   Method |       Mean |     Error |    StdDev | Ratio | Allocated |
|--------- |-----------:|----------:|----------:|------:|----------:|
|  Current | 36.5179 ns | 0.7490 ns | 0.7356 ns |  1.00 |         - |
| Improved |  0.6491 ns | 0.0549 ns | 0.0675 ns |  0.02 |         - |

After

|   Method |      Mean |     Error |    StdDev |    Median | Ratio | RatioSD | Allocated |
|--------- |----------:|----------:|----------:|----------:|------:|--------:|----------:|
|  Current | 2.6616 ns | 0.0949 ns | 0.1939 ns | 2.5941 ns |  1.00 |    0.00 |         - |
| Improved | 0.8948 ns | 0.0588 ns | 0.1487 ns | 0.8456 ns |  0.35 |    0.06 |         - |

### BitwiseXorBenchmark

Before

|   Method |       Mean |     Error |    StdDev | Ratio | Allocated |
|--------- |-----------:|----------:|----------:|------:|----------:|
|  Current | 36.6601 ns | 0.7728 ns | 1.2911 ns |  1.00 |         - |
| Improved |  0.9368 ns | 0.0571 ns | 0.1389 ns |  0.02 |         - |

After

|   Method |      Mean |     Error |    StdDev | Ratio | RatioSD | Allocated |
|--------- |----------:|----------:|----------:|------:|--------:|----------:|
|  Current | 2.7754 ns | 0.0955 ns | 0.1369 ns |  1.00 |    0.00 |         - |
| Improved | 0.7530 ns | 0.0529 ns | 0.0495 ns |  0.27 |    0.02 |         - |